fix: replace int by double for weights & biases

+++ src/neural/xor/config_manager.h
@@ -2,12 +2,12 @@
 #define DEFAULT_CONFIG_FILE "default_xor.nnconf"
 
 void write_config(const int nb_inputs, const int nb_layers,
-                  const int* nb_nodes_p, const int** biases_pp,
-                  const int*** weights_ppp)
+                  const int* nb_nodes_p, const double** biases_pp,
+                  const double*** weights_ppp)
 
 int get_nb_layers();
 
 void get_nb_nodes(const int nb_layers, int* nb_nodes_p);
 
 void get_config(const int nb_inputs, const int nb_layers, const int* nb_nodes_p,
-                int** biases_pp, int*** weights_ppp);
+                double** biases_pp, double*** weights_ppp);
